refactor: move components to formly folder and tslint rule
