+++ src/app/pages/registration/formly/registration-heading-field/registration-heading-field.component.spec.ts
