+++ src/app/pages/registration/registration-page.module.ts
@@ -4,7 +4,7 @@ import { FormlyModule } from '@ngx-formly/core';
 
 import { SharedModule } from 'ish-shared/shared.module';
 
-import { RegistrationAddressFieldComponent } from './registration-address-field/registration-address-field.component';
+import { RegistrationAddressFieldComponent } from './formly/registration-address-field/registration-address-field.component';
 import { RegistrationHeadingFieldComponent } from './registration-heading-field/registration-heading-field.component';
 import { RegistrationPageComponent } from './registration-page.component';
 import { RegistrationTacFieldComponent } from './registration-tac-field/registration-tac-field.component';
