+++ tslint.json
@@ -587,7 +587,9 @@
           "^.*/src/app/shared/[a-z][a-z0-9-]+/components/([a-z][a-z0-9-]+)/\\1\\.component\\.ts$",
           "^.*/projects/[a-z][a-z0-9-]+/src/app/components/([a-z][a-z0-9-]+)/\\1\\.component\\.ts$",
           "^.*/src/app/shared/address-forms/components/([a-z][a-z0-9-]+)/\\1\\.factory\\.ts$",
+          // formly
           "^.*/src/app/shared/formly/(components|wrappers|types|extensions|utils|dev)/.*$",
+          "^.*/src/app/pages/([a-z][a-z0-9-]+)/formly/.*$",
           // identity providers
           "^.*/src/app/core/identity-provider/.*$",
           // aggregation modules
